[32/64-BIT] - Firing up with CLI
Both 32 and 64-bit are firing up - MMU is working, process is installing the initial CLI binary.
There is nothing more than this - it stops at the "*" symbol for now, but it does show the 64-bit side
of things in particular is starting to align with the 32-bit build.

Output currently looks like this:

RASPBERRY PI 3 MODEL B+ 948MB 64-bit
IMP OS 0.2 (Jul 19 2019 15:17:49)
[MMU : ENABLED]
SD CARD: 14940MB (SD16G) [FAT32]

*

+++ kernel/c/common/swi.c
@@ -48,11 +48,11 @@ swi_function handler[] =
  * Dispatches the relevant internal SWI function, based on the supplied SWI number.
  *
  */
-unsigned int _kernel_swi_handler( unsigned int swi )
+unsigned int _kernel_swi_handler( uintptr_t swi )
 {
 
 	// notify the screen of the SWI number invoked, as well as the value of in.r[0]
-	unsigned int *regs_ptr;
+	uintptr_t *regs_ptr;
 	regs_ptr = (uintptr_t *) swi_params_ptr_in;
 	// _kernel_video_print_string( "_kernel_swi_handler : " ); _kernel_video_print_hex( swi         ); _kernel_video_print_string( "\n" );
 	// _kernel_video_print_string( "in.r[0] = 0x"           ); _kernel_video_print_hex( regs_ptr[0] ); _kernel_video_print_string( "\n" );
